Move and rename all of the no results stuff.
